pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "research-agent-harness"
version = "0.1.0"
description = "Autonomous research agent: from topic to submission-ready paper."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "agentic-programming",
]

[project.scripts]
research-harness = "research_harness.main:main"

[tool.setuptools.packages.find]
include = ["research_harness*"]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests that call a real LLM (slow, needs API key or claude-code)",
]

[project.urls]
Homepage = "https://github.com/Fzkuji/Research-Agent-Harness"
